Avoid rounding errors by casting display height/width to float when calculating the multipliers for subtitle positioning.
git-svn-id: http://svn.mythtv.org/svn/trunk@22585 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/NuppelVideoPlayer.cpp
@@ -7205,8 +7205,8 @@ void NuppelVideoPlayer::DisplayAVSubtitles(void)
                     }
                 }
 
-                float hmult = osd->GetSubtitleBounds().width() / video_disp_dim.width();
-                float vmult = osd->GetSubtitleBounds().height() / video_disp_dim.height();
+                float hmult = osd->GetSubtitleBounds().width() / (float) video_disp_dim.width();
+                float vmult = osd->GetSubtitleBounds().height() / (float) video_disp_dim.height();
 
                 rect->x = (int)(rect->x * hmult);
                 rect->y = (int)(rect->y * vmult);
